fix: got user create event to notify other services consistently
