+++ example.env
@@ -50,4 +50,4 @@ USER_NATIONAL_ID=000000
 
 
 # Other Service's API Keys
-EVENT_API_KEY=IKZ)EeG^OBea}i4K?T=J`p+8vP@#R+5W(rR%?GvD{|QT3
+API_KEY=IKZ)EeG^OBea}i4K?T=J`p+8vP@#R+5W(rR%?GvD{|QT3
